Included org-link-abbrev-alist
The #+LINK notation was removed from the heading of all org buffers and
org-link-abbrev was defined in general-org.el (for global, device
independent links) and added to in linux-org.el, osx-org.el, and
windows-org.el for device dependent links.

init.el was modified to ensure that general-org.el loads before any of
the device dependent files.