Update windows-org.el
There was a problem with "U:\" which I realized a while ago--emacs was not loading anything past windows-org.el with an "end of line" error. Apparently the \" is used as an escape character unless it's followed by something else (such as a directory name in the case of "U:\org". I changed it to a double slash and everything seems to be working fine on Windows now.