U: general-org, agenda-custom, and capture-temp...
general-org.el:
- Turned off property inheritance because it’s too complex and
unnecessary for my purposes.
- Added variables to re-file items with C-c C-w based on path
(ref.org/Heading/Sub-Heading/Sub-Sub-Heading/etc.), which also includes
auto-completion.

agenda-custom.el
- Excluded SCHEDULED items by default; they no longer appear in the
timeline agenda created with C-a a.

capture-templates.el
- Added “Tasks from meeting:” to the end of the meeting template; from
now on tasks from meeting with be re-filed to their correct area but
maintain a link in the meeting; the tasks will also have a link to the
meeting in the :Note: property.