Update to osx.el and platform specific files
Updates to osx.el are specific to OSX and include:

1) A commented-out full screen capability which, upon opening emacs,
switches it to a different workspace and makes it full screen;

2) Code to maximize Emacs by default upon start up. This is done by
setting the width and height which are dependent on character size.
This works for a 13” Retina MacBook Pro.

Updates to linux-org.el, osx-org.el, and windows-org.el include:

1) Changing (add-to-list ‘org-link-abbrev-alist …) to multiple
add-to-lists. On OSX having a single add-to-list was not working. This
functionality has not been tested in linux-org.el yet. A single list
was working in window-org.el, but it was changed to multiple for
consistency.