Update to capture-templates.el and general.el
The shopping capture template was creating a shopping headline in
ref.org because it originally was, when not given a home, re-filing to
fin.org:Shopping. It now, when not given a home, defaults to ref.org
just like everything else.

Commented out code has also been added to general.el to deal with file
encoding, something I didn't even embarrassingly know was a problem.