Update to general-org.el
Tried to fix my/multimedia-capture to work with multiple APIs but I just
don't think it's going to happen. I spend the new few hours trying to
get chrome to properly interface with JSON objects using bookmarklets,
and they worked on some things. It seems the harder I try, the further
away I get from the solution. In the meantime, the current solution
seems to work just fine. The only potential improvement would be to have
it automatically load the reference URL.