Update to org-protocol-templates.el and windows.el
Enabled org-protocol-capture-readability to accept prefix arguments in
the form C-0. I only really use C-0 to move the capture to the current
cursor location.

Finally figured out how to load https pages from windows. It required me
to have a version of openssl (which ships with Git). I was able to use
the correct file path and now it works.