Update to org-capture-templates.el
Added automatic prompting for certain properties in the contact and
business templates. In order to reuse fields (e.g. %\\1), the backslash
needs to be escaped. It’s also important to note that date prompts
can’t be reused and don’t count as a number which is unfortunate.

The date of Christmas will have to be updated each year.

More information on capture templates can be found here:
http://orgmode.org/manual/Template-expansion.html.