Update to org-capture-templates.el
Added :empty-lines property to each capture template. It ensures an
empty line is added to the beginning and end of each capture template.
Quick testing found that even if there were multiple line breaks in the
target location, filing the item would result in a single line above and
a single line below.

The org-id for this task is:
18ee4d42-2817-4755-a090-5269d8cede17