Update to osx.el
Set visible-bell to nil (which is the default) and set
ring-bell-function to ‘ignore so no sound is made. This is a
work-around to prevent a pop-up box in the middle of Emacs when the
bell would usually go off.

This was complete for org-id FB2AA049-A030-444F-ACD4-7CC3C69D8752.