Update to org-general.el
Advised dired-rename-file to automatically add a link to the new
location of every file rename in dired. This is useful for, say, filing
receipts that I intend to link to in a capture template. The org-links
can be inserted with org-insert-link.

Completed for task org-id 5301FAF1-9803-4F8C-97F7-E6E2EEE89578.